+++ crates/objc2/src/__macro_helpers/mod.rs
@@ -35,7 +35,7 @@ pub use self::common_selectors::{alloc_sel, dealloc_sel, init_sel, new_sel};
 pub use self::convert::{ConvertArgument, ConvertArguments, ConvertReturn, TupleExtender};
 pub use self::declare_class::{
     ClassBuilderHelper, ClassProtocolMethodsBuilder, MaybeOptionRetained, MessageReceiveRetained,
-    RetainedReturnValue,
+    RetainedReturnValue, ThreadKindAutoTraits,
 };
 pub use self::declared_ivars::DeclaredIvarsHelper;
 pub use self::image_info::ImageInfo;