+++ crates/objc2/src/macros/extern_class.rs
@@ -314,6 +314,8 @@ macro_rules! __extern_class_inner {
             }
 
             const __INNER: () = ();
+
+            type __SubclassingType = Self;
         }
 
         $($attr_impl)*